tests/end_to_end/requirements.txt
paho-mqtt
kpn-senml
requests
websockets